refactor(storage): export IndexSeriesCursor
This allows IDPE to use the same implementation, rather than duplicate
code.
Also copied unit tests from IDPE.
