+++ storage/readservice/store.go
@@ -9,28 +9,18 @@ import (
 	"github.com/influxdata/influxdb"
 	"github.com/influxdata/influxdb/kit/tracing"
 	"github.com/influxdata/influxdb/models"
-	"github.com/influxdata/influxdb/storage"
 	"github.com/influxdata/influxdb/storage/reads"
 	"github.com/influxdata/influxdb/storage/reads/datatypes"
-	"github.com/influxdata/influxdb/tsdb"
 	"github.com/influxdata/influxdb/tsdb/cursors"
 	"github.com/influxdata/influxql"
 )
 
-// Viewer is used by the store to query data from time-series files.
-type Viewer interface {
-	CreateCursorIterator(ctx context.Context) (tsdb.CursorIterator, error)
-	CreateSeriesCursor(ctx context.Context, orgID, bucketID influxdb.ID, cond influxql.Expr) (storage.SeriesCursor, error)
-	TagKeys(ctx context.Context, orgID, bucketID influxdb.ID, start, end int64, predicate influxql.Expr) (cursors.StringIterator, error)
-	TagValues(ctx context.Context, orgID, bucketID influxdb.ID, tagKey string, start, end int64, predicate influxql.Expr) (cursors.StringIterator, error)
-}
-
 type store struct {
-	viewer Viewer
+	viewer reads.Viewer
 }
 
 // NewStore creates a store used to query time-series data.
-func NewStore(viewer Viewer) reads.Store {
+func NewStore(viewer reads.Viewer) reads.Store {
 	return &store{viewer: viewer}
 }
 
@@ -48,7 +38,7 @@ func (s *store) ReadFilter(ctx context.Context, req *datatypes.ReadFilterRequest
 	}
 
 	var seriesCursor reads.SeriesCursor
-	if seriesCursor, err = newIndexSeriesCursor(ctx, orgID, bucketID, req.Predicate, s.viewer); err != nil {
+	if seriesCursor, err = reads.NewIndexSeriesCursor(ctx, orgID, bucketID, req.Predicate, s.viewer); err != nil {
 		return nil, tracing.LogError(span, err)
 	} else if seriesCursor == nil {
 		return nil, nil
@@ -71,7 +61,7 @@ func (s *store) ReadGroup(ctx context.Context, req *datatypes.ReadGroupRequest) 
 	}
 
 	newCursor := func() (reads.SeriesCursor, error) {
-		return newIndexSeriesCursor(ctx, orgID, bucketID, req.Predicate, s.viewer)
+		return reads.NewIndexSeriesCursor(ctx, orgID, bucketID, req.Predicate, s.viewer)
 	}
 
 	return reads.NewGroupResultSet(ctx, req, newCursor), nil
